HomeWork

+++ src/lesson_19/HomeWork/Calculator.java
@@ -25,7 +25,7 @@ public class Calculator {
     }
 
     // Метод "деление"
-    public static double division(int a, int b) {
-        return (double) a / b;
+    public static double division(double a, double b) {
+        return a / b;
     }
 }
